Remove jenkins swarm slaves from site.pp
In the following days we need to redeploy a bunch of jenkins
slaves for product-ci and old-stable-ci and none of them will use
jenkins-swarm-slave. Removing this definitions will make using new
hiera roles for those slaves a lot easier and if we ever need to
deploy another swarm slave - we can just use product_swarm_slave role.

Change-Id: Ifc38506010aeefa5b5c6701e846e8353ec7e82a9