Updated version of puppetlabs-firewall module
Version of 'puppetlabs-firewall' module updated from 1.1.3 to 1.7.1.
It needs to have an ability to create a firewall rule with
inversion.

To fix the bug we should use 1.7.1 version of the module because 1.1.3 version
couldn't implement a firewall rule with inversion. For example, the rule:
-A FORWARD -i docker0 ! -o docker0 -m comment --comment "One more docker rule" -j ACCEPT.

Old version of the module is used non-correct pattern to parse such rules,
that's why old version is doesn't works.

Change-Id: Id50b92e8c9a998fad439bf22db624155146e6c15
Related-bug: #1496399