Use libvirt 1.2.12 on RHEL bases distros
Since we plan to use libvirt 1.2.12 on all related distros
(Ubuntu,RHEL), its required to use versionlocking to pin
to specific pkg version.

Also remove the FIXME tmp code as the custom rule for polkit is
provided by the libvirt-daemon 1.2.12 package.

Add jenkins user to the kvm group as kvm group is allowed in the
polkit's rule, which comes from libvirt-daemon v1.2.12.

Change-Id: I8e5d4a085c3bcf0659865c161f45bba52178876e