Put credentials into dedicated file
In nodepool, for security, the sensitive information like passwords for
connecting to a database should not be stored in project-config
repo, which stores nodepool.yaml cfg file. The dburi, providers or
targets user names, passwords, api keys should be "appended" to nodepool.yaml
after the project-config repo clone has happened, as an additional step.

Also, remove the vcsrepo usage - replace it with executing
the "project_config_sync.sh" script, which will do the following:
- clone the project-config repo,
- set proper dburi, providers and targets credentials in nodepool.yaml.

Change-Id: Iaeff27f4904f3d69c2d18b26684d774de7ced861