New script to get hiera from Gerrit to a puppet master added
New script to get hiera from Gerrit to a puppet master added, also
"fuel_project::puppet::master" module is updated to create the script and
configure all data for it.

This script allows to use hiera stored in Gerrit instead storing it
as a local git repo on a puppet master.

Change-Id: I116e452a1a13d339227b0e3ca3ee7b7d4f5b2382