Version of the "elasticsearch" pinned to '1.6.2' version
The part of the code to pin the "elasticsearch" package version
to 1.6.2 added.
This part is needed because of the latest version os the "elasticsearch"
package doesn't work with current version of the "fuel-stats-analytics"
package, because an ability to execute queries with the dot symbol
in a bode was removed from the "elasticsearch" package.

Each run of the "apt-get upgrade" command installs the latest
version of the "elasticsearch" package, but we want to have
'1.6.2' version there.

Change-Id: Iba743020c660ef92d856502c5887f1dd5e705674